--- app/src/main/java/com/eatssu/android/data/model/Dodam.kt
@@ -1,3 +1,0 @@
-package com.eatssu.android.data.model
-
-data class Dodam(val menu: String, val price: String, val rate: Double)
