--- app/src/main/java/com/eatssu/android/data/model/Haksik.kt
@@ -1,8 +1,0 @@
-package com.eatssu.android.data.model
-
-
-public class Haksik(
-    val menu: String,
-    val price: String,
-    val rate: Double
-    )
