--- app/src/main/java/com/eatssu/android/data/model/Kitchen.kt
@@ -1,3 +1,0 @@
-package com.eatssu.android.data.model
-
-data class Kitchen(val menu: String, val price: String, val rate: Double)
