--- app/src/main/java/com/eatssu/android/data/model/Snack.kt
@@ -1,3 +1,0 @@
-package com.eatssu.android.data.model
-
-data class Snack(val menu: String, val price: String, val rate: Double)
