--- app/src/main/java/com/eatssu/android/data/model/review.kt
@@ -1,9 +1,0 @@
-package com.eatssu.android.data.model
-
-import java.net.IDN
-
-data class review(
-    val id : String,
-    val content : String,
-    val date : String,
-)
