+++ app/src/main/java/com/eatssu/android/ui/main/BreakfastFragment.kt
@@ -2,31 +2,19 @@ package com.eatssu.android.ui.main
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.LinearLayout
-import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.eatssu.android.R
-import com.eatssu.android.RestaurantType
+import com.eatssu.android.data.enums.RestaurantType
 import com.eatssu.android.data.RetrofitImpl
-import com.eatssu.android.data.model.*
 import com.eatssu.android.data.model.response.GetMenuInfoListResponse
-import com.eatssu.android.data.model.response.MenuBaseResponse
 import com.eatssu.android.data.service.MenuService
-import com.eatssu.android.data.service.RetrofitInterface
 import com.eatssu.android.databinding.FragmentBreakfastBinding
-import com.eatssu.android.databinding.FragmentDinnerBinding
-import com.eatssu.android.ui.calendar.CalendarData
 import com.eatssu.android.ui.infopage.*
-import com.eatssu.android.ui.menuadapter.*
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import com.eatssu.android.ui.menuAdapter.*
 import retrofit2.Retrofit
 
 class BreakfastFragment : Fragment() {
