+++ app/src/main/java/com/eatssu/android/ui/main/DinnerFragment.kt
@@ -7,19 +7,15 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.LinearLayout
-import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.eatssu.android.RestaurantType
+import com.eatssu.android.data.enums.RestaurantType
 import com.eatssu.android.data.RetrofitImpl
 import com.eatssu.android.data.model.response.GetMenuInfoListResponse
 import com.eatssu.android.data.service.MenuService
 import com.eatssu.android.databinding.FragmentDinnerBinding
-import com.eatssu.android.databinding.FragmentLunchBinding
 import com.eatssu.android.ui.infopage.*
-import com.eatssu.android.ui.MenuAdapter
-import com.eatssu.android.ui.menuadapter.*
+import com.eatssu.android.ui.menuAdapter.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
