+++ app/src/main/java/com/eatssu/android/ui/menuAdapter/DodamAdapter.kt
@@ -1,26 +1,19 @@
-package com.eatssu.android.ui.menuadapter
+package com.eatssu.android.ui.menuAdapter
 
 import android.content.Intent
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import com.eatssu.android.R
 import com.eatssu.android.data.model.response.GetMenuInfoListResponse
-import com.eatssu.android.data.model.response.MenuBaseResponse
 import com.eatssu.android.databinding.ItemDodamBinding
-import com.eatssu.android.databinding.ItemFoodBinding
-import com.eatssu.android.databinding.ItemGisikBinding
-import com.eatssu.android.databinding.ItemSnackBinding
 import com.eatssu.android.ui.review.ReviewListActivity
 
 
-class GisikAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>) :
-    RecyclerView.Adapter<GisikAdapter.ViewHolder>() {
+class DodamAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>):
+    RecyclerView.Adapter<DodamAdapter.ViewHolder>() {
 
-    inner class ViewHolder(private val binding: ItemGisikBinding) :
+    inner class ViewHolder(private val binding: ItemDodamBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind(position: Int) {
@@ -32,7 +25,7 @@ class GisikAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding =
-            ItemGisikBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+            ItemDodamBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
