+++ app/src/main/java/com/eatssu/android/ui/menuAdapter/GisikAdapter.kt
@@ -1,27 +1,19 @@
-package com.eatssu.android.ui.menuadapter
+package com.eatssu.android.ui.menuAdapter
 
 import android.content.Intent
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import com.eatssu.android.R
 import com.eatssu.android.data.model.response.GetMenuInfoListResponse
-import com.eatssu.android.data.model.response.MenuBaseResponse
-import com.eatssu.android.databinding.ItemDodamBinding
-import com.eatssu.android.databinding.ItemFoodBinding
 import com.eatssu.android.databinding.ItemGisikBinding
-import com.eatssu.android.databinding.ItemHaksikBinding
-import com.eatssu.android.databinding.ItemSnackBinding
 import com.eatssu.android.ui.review.ReviewListActivity
 
 
-class HaksikAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>):
-    RecyclerView.Adapter<HaksikAdapter.ViewHolder>() {
+class GisikAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>) :
+    RecyclerView.Adapter<GisikAdapter.ViewHolder>() {
 
-    inner class ViewHolder(private val binding: ItemHaksikBinding) :
+    inner class ViewHolder(private val binding: ItemGisikBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind(position: Int) {
@@ -33,7 +25,7 @@ class HaksikAdapter(private val dataList: List<GetMenuInfoListResponse.MenuInfo>
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding =
-            ItemHaksikBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+            ItemGisikBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
